更新代码以支持64位寄存器和数据处理，完全跑通64测试！
主要改动包括：
- 将浮点寄存器和程序计数器的类型从32位更改为64位，确保与RV64I架构兼容。
- 移除不再使用的ALU相关文件，简化代码结构。
- 更新调试输出格式，统一使用64位格式化，提升可读性。
- 修改相关测试用例以适应64位寄存器的使用。

+++ include/common/cpu_interface.h
@@ -26,8 +26,8 @@ public:
     virtual void setRegister(RegNum reg, uint64_t value) = 0;
     
     // 浮点寄存器访问
-    virtual uint32_t getFPRegister(RegNum reg) const = 0;
-    virtual void setFPRegister(RegNum reg, uint32_t value) = 0;
+    virtual uint64_t getFPRegister(RegNum reg) const = 0;
+    virtual void setFPRegister(RegNum reg, uint64_t value) = 0;
     virtual float getFPRegisterFloat(RegNum reg) const = 0;
     virtual void setFPRegisterFloat(RegNum reg, float value) = 0;
